// web/src/lib/api/generated/models/internalApiHandlersUserSettingResponse.ts
/**
 * Generated by orval v7.16.0 🍺
 * Do not edit manually.
 * ViewRA Media Server API
 * Self-hosted media server for movies, TV shows, and music
 * OpenAPI spec version: 0.0.1
 */

export interface InternalApiHandlersUserSettingResponse {
  key?: string
  updated_at?: string
  value?: string
}